+++ views/partials/head.html
@@ -1,30 +1,24 @@
 {{define "head"}}
 <meta charset="UTF-8" />
-<!-- <link rel="apple-touch-icon" sizes="76x76" href="/assets/img/apple-icon.png" /> -->
-<!-- <link rel="icon" type="image/png" href="/assets/img/favicon.png" /> -->
 
-  <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
 <title>{{ .Title }}</title>
-<!--     Fonts and icons     -->
-<link href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet" />
 <!-- Font Awesome Icons -->
+<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css"
   integrity="sha512-z3gLpd7yknf1YoNbCzqRKc4qyor8gaKU1qmn+CShxbuBusANI9QpRohGBreCFkKxLhei6S9CQXFEbbKuqLg0DA=="
   crossorigin="anonymous" referrerpolicy="no-referrer" />
-<!-- Nucleo Icons -->
-<link href="/assets/css/nucleo-icons.css" rel="stylesheet" />
-<link href="/assets/css/nucleo-svg.css" rel="stylesheet" />
-<!-- Popper -->
 <!-- Main Styling -->
-  <!-- <script src="https://cdn.tailwindcss.com"></script> -->
-<link href="/assets/css/argon-dashboard-tailwind.css?v=1.0.1" rel="stylesheet" />
-<link href="/assets/css/styles.css" rel="stylesheet" />
+<script src="https://cdn.tailwindcss.com"></script>
 <script src="https://unpkg.com/htmx.org@1.9.5"></script>
 <script src="https://unpkg.com/@popperjs/core@2"></script>
 <script src="https://unpkg.com/htmx.org/dist/ext/class-tools.js"></script>
 <script src="https://unpkg.com/hyperscript.org@0.9.11"></script>
-<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+<script>
+  tailwind.config = {
+    darkMode: 'class',
+  }
+</script>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
 {{end}}
