+++ views/partials/header.html
@@ -1,4 +1,3 @@
-{{define "header"}}
 <div class="bg-gray-200 dark:bg-gray-800">
     <div class="container flex items-center px-6 py-4 mx-auto overflow-x-auto whitespace-nowrap">
         <a href="#" class="text-gray-600 dark:text-gray-200">
@@ -51,5 +50,3 @@
         </a>
     </div>
 </div>
-
-{{end}}
